Remove memory and table layout information from emscripten_metadata section
We are moving away from the world where the memory layout if
decided by the python driver code and towards a world where
the wasm-ld decides the final layout.

For relocatable module (SIDE_MODULE and MAIN_MODULE) the size
and alignment of memory and table segments is already encoded
in the `dylink` sections.

For non-relocatable modules I don't think this information is
needed by the embedded as they can simply look at the requirements
on the table and memory that a given module imports.

In general symbols like `__global_base` or `__data_end` and
`__heap_base` are details we don't export by default but can
be explictly exported if needed.